# apps/docs/api-reference/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SnipShare API",
    "description": "Revolutionize your coding experience with SnipShare - the all-in-one platform for creating, executing, and sharing Node.js code snippets.",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/sessions/github": {
      "post": {
        "summary": "Authenticate with GitHub",
        "tags": ["auth"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string"
                  }
                },
                "required": ["code"],
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    }
                  },
                  "required": ["token"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/organizations/{slug}": {
      "get": {
        "summary": "Get details of an organization by slug",
        "tags": ["organizations"],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "slug",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "organization": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "slug": {
                          "type": "string"
                        },
                        "domain": {
                          "type": "string",
                          "nullable": true
                        },
                        "shouldAttachUsersByDomain": {
                          "type": "boolean"
                        },
                        "avatarUrl": {
                          "type": "string",
                          "format": "uri",
                          "nullable": true
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "ownerId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "slug",
                        "domain",
                        "shouldAttachUsersByDomain",
                        "avatarUrl",
                        "createdAt",
                        "updatedAt",
                        "ownerId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": ["organization"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update an organization",
        "tags": ["organizations"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "domain": {
                    "type": "string",
                    "nullable": true
                  },
                  "shouldAttachUsersByDomain": {
                    "type": "boolean"
                  }
                },
                "required": ["name"],
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "slug",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "enum": ["null"],
                  "nullable": true
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Shutdown an organization",
        "tags": ["organizations"],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "slug",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "enum": ["null"],
                  "nullable": true
                }
              }
            }
          }
        }
      }
    },
    "/organizations/{slug}/owner": {
      "patch": {
        "summary": "Transfer an organization ownership",
        "tags": ["organizations"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "transferToUserId": {
                    "type": "string"
                  }
                },
                "required": ["transferToUserId"],
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "slug",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "enum": ["null"],
                  "nullable": true
                }
              }
            }
          }
        }
      }
    },
...
  }
}